refactor(commands): relay 测试按域下沉到各领域模块 + 共享 test_support
4900 行的 tests.rs（拆分时有意保留的整块跨域集成测试）按「测试引用哪个
领域模块的条目」归位到各自领域文件的 mod tests；被两个以上域引用的
脚手架（凭证构造/mock 服务/seed helper/model_verification 类型导入/
ResetVerifier）收进 test_support.rs（cfg(test) 模块，pub(crate) 再导出
原顶部公共 use）。分类脚本 + 编译修复循环完成，测试集合不变（3819 通过）。
