+++ src-tauri/src/commands/relay/windows.rs
@@ -284,3 +284,473 @@ async fn ensure_token_outlasts_a_payment<R: tauri::Runtime>(
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::commands::relay::test_support::*;
+
+    /// ⭐ 回归闸：sub2api 充值页必须打开**签名配置的 URL**，不再读站点公开设置的
+    /// 支付开关去猜 `/purchase` 还是 `/redeem`。
+    ///
+    /// 三个断言互相补充：
+    /// 1. 请求日志**不含** `/api/v1/settings/public` —— 路由事实已归签名目录，
+    ///    生产充值路径不该再读站点公开设置。
+    /// 2. 请求日志恰好是「开窗前续期 + 取账号档案」两个请求 —— 证明 token 寿命续期
+    ///    与登录态注入这些既有行为没有被这次改动顺带丢掉。
+    /// 3. 窗口打开的 URL **逐字符等于**配置值 —— 配置里故意用了不可推导的路径
+    ///    （`/topup-center?flow=card`），推导逻辑造不出它。
+    ///
+    /// 用 middleware 记录**所有**请求的 path（含未注册路由的 404）：逐 handler 记录会漏掉
+    /// 「代码打了但我们没 serve 的路径」，那种漏记正好把要抓的回归放跑。
+    #[tokio::test]
+    async fn sub2api_purchase_uses_signed_url() {
+        use axum::{
+            extract::Request, middleware, middleware::Next, routing::get, routing::post, Json,
+            Router,
+        };
+
+        let requests = Arc::new(Mutex::new(Vec::<String>::new()));
+        let every_request = Arc::clone(&requests);
+
+        let app = Router::new()
+            // `ensure_token_outlasts_a_payment` 开窗前无条件续期要打的端点。
+            .route(
+                "/api/v1/auth/refresh",
+                post(|| async move {
+                    Json(serde_json::json!({
+                        "code": 0,
+                        "message": "success",
+                        "data": {
+                            "access_token": "fresh-access-token",
+                            "refresh_token": "fresh-refresh-token",
+                            "expires_at": 4_102_444_800_000_i64
+                        }
+                    }))
+                }),
+            )
+            // `auth_user_from_profile` 要的账号档案（信封 `data` 里必须有 `id`）。
+            .route(
+                "/api/v1/user/profile",
+                get(|| async move {
+                    Json(serde_json::json!({
+                        "code": 0,
+                        "message": "success",
+                        "data": { "id": 7, "email": "saved-account", "username": "Saved Account" }
+                    }))
+                }),
+            )
+            // ⚠️ 陷阱端点：旧实现靠它读站点公开设置的支付开关猜路由。这里故意把它
+            // 配成一个能正常解析的 sub2api 响应 —— 只要充值流程还来问它，下面的
+            // 断言当场红。
+            .route(
+                "/api/v1/settings/public",
+                get(|| async move { Json(sub2api_discovery_body()) }),
+            )
+            .layer(middleware::from_fn(move |req: Request, next: Next| {
+                let requests = Arc::clone(&every_request);
+                async move {
+                    requests.lock().unwrap().push(req.uri().path().to_string());
+                    next.run(req).await
+                }
+            }));
+        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
+            .await
+            .expect("bind purchase test server");
+        let addr = listener.local_addr().expect("server address");
+        let server = tokio::spawn(async move {
+            axum::serve(listener, app)
+                .await
+                .expect("serve purchase app");
+        });
+
+        // relay 行存 http origin（测试站是本机 http 服务）；签名目录的 `purchase_url`
+        // 用同 host:port 的 https 形态 —— `normalize_site_origin` 强制 https 且保留端口，
+        // 所以两者恰好同源、`configured_purchase_url` 解析成功。
+        let (app, relay_id) =
+            saved_relay_app(&format!("http://{addr}"), discovery::BackendKind::Sub2Api);
+        let op = relay_credentials(&app, relay_id);
+
+        let configured = format!("https://{addr}/topup-center?flow=card");
+        let config = remote_config::RemoteConfig {
+            relay_directory: remote_config::RelayDirectoryPolicy {
+                blocked_hosts: vec![],
+                sites: std::collections::BTreeMap::from([(
+                    "127.0.0.1".to_string(),
+                    remote_config::RelayDirectorySite {
+                        veridrop_host: None,
+                        entry_url: None,
+                        purchase_url: Some(configured.clone()),
+                        usage_url: None,
+                        display_name: None,
+                    },
+                )]),
+            },
+            ..remote_config::RemoteConfig::default()
+        };
+        let purchase_url = remote_config::configured_purchase_url(&config, &op.site_origin)
+            .expect("签名目录解析不该报错")
+            .expect("这个站在目录里配了购买入口");
+
+        let window = purchase::purchase_window(relay_id, &op.site_origin);
+        open_sub2api_site_window(app.handle(), op, window, purchase_url)
+            .await
+            .expect("开充值窗");
+
+        let paths = requests.lock().unwrap().clone();
+        assert_eq!(
+            paths,
+            vec!["/api/v1/auth/refresh", "/api/v1/user/profile"],
+            "充值流程只该打「开窗前续期 + 取账号档案」两个请求；\
+                 出现 /api/v1/settings/public 说明又回去按公开设置猜路由了"
+        );
+
+        let window = app
+            .get_webview_window(&purchase::window_label(relay_id))
+            .expect("充值窗应该开出来了");
+        let opened = window
+            .url()
+            .expect("mock 窗口能读回创建时的 URL")
+            .to_string();
+        assert_eq!(
+            opened, configured,
+            "打开的外部 URL 必须恰好是签名配置值，不是推导出的 /purchase 或 /redeem"
+        );
+
+        server.abort();
+    }
+
+    // ======================================================================
+    // NewAPI 充值分派（Task 7）。全部直接驱动 `dispatch_purchase` —— 生产
+    // `load_cached()` 用生产公钥验签，测试无法（也不该）伪造缓存，这个接缝与
+    // `open_sub2api_purchase_window` 的「参数化只为可测」是同一惯例。
+    // 协议字面量一律从 `newapi` owner 派生（backend.rs 的架构闸钉着）。
+    // ======================================================================
+
+    /// 起一个「记录全部请求 path」的哨兵站点：只应答 NewAPI 探测端点的形状
+    /// （`/api/status`），其余 404 —— 任何路径都会进日志（middleware 不挑路由）。
+    async fn recording_newapi_sentinel(
+    ) -> (String, Arc<Mutex<Vec<String>>>, tokio::task::JoinHandle<()>) {
+        use axum::{extract::Request, middleware, middleware::Next, routing::get, Json, Router};
+
+        let requests = Arc::new(Mutex::new(Vec::<String>::new()));
+        let every_request = Arc::clone(&requests);
+        let app = Router::new()
+            .route(
+                "/api/status",
+                get(|| async move { Json(newapi_discovery_body()) }),
+            )
+            .layer(middleware::from_fn(move |req: Request, next: Next| {
+                let requests = Arc::clone(&every_request);
+                async move {
+                    requests.lock().unwrap().push(req.uri().path().to_string());
+                    next.run(req).await
+                }
+            }));
+        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
+            .await
+            .expect("bind newapi purchase sentinel");
+        let addr = listener.local_addr().expect("sentinel address");
+        let server = tokio::spawn(async move {
+            axum::serve(listener, app)
+                .await
+                .expect("serve newapi purchase sentinel");
+        });
+        (format!("http://{addr}"), requests, server)
+    }
+
+    /// 等 monitor 任务收场把 lease 还回去（`open` 返回与后台任务 drop lease 之间
+    /// 有毫秒级竞态，轮询等待而不是睡固定时长）。
+    async fn until_purchase_lease_released(
+        app: &tauri::App<tauri::test::MockRuntime>,
+        relay_id: i64,
+    ) {
+        let coordinator = Arc::clone(&app.state::<AppState>().purchase_sessions);
+        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
+        while coordinator.is_active(relay_id) {
+            assert!(
+                std::time::Instant::now() < deadline,
+                "命令返回后 lease 必须随即释放"
+            );
+            tokio::time::sleep(std::time::Duration::from_millis(5)).await;
+        }
+    }
+
+    /// 把 `saved_relay_app` 存好的 refresh credential 覆写成空白（模拟凭据缺失的行）。
+    fn blank_saved_refresh_credential(app: &tauri::App<tauri::test::MockRuntime>, relay_id: i64) {
+        let state = app.state::<AppState>();
+        let conn = state.db.conn.lock().expect("lock memory database");
+        conn.execute(
+            "UPDATE loongport_relay SET refresh_token = ?1 WHERE id = ?2",
+            rusqlite::params!["   ", relay_id],
+        )
+        .expect("blank refresh credential");
+    }
+
+    #[tokio::test]
+    async fn newapi_purchase_dispatch_times_out_without_touching_sub2api_endpoints() {
+        let (origin, requests, server) = recording_newapi_sentinel().await;
+        // relay 行存 https 形态的哨兵 origin（生产行的 origin 在导入时就归一成 https），
+        // purchase_url 用同 host:port 的 https 形态 —— `configured_purchase_url` 才解析得出。
+        let (app, relay_id) = saved_relay_app(
+            &origin.replacen("http://", "https://", 1),
+            discovery::BackendKind::NewApi,
+        );
+        let op = relay_credentials(&app, relay_id);
+        assert_eq!(
+            op.refresh_token.as_deref(),
+            Some("saved-refresh-token"),
+            "前提：这行有非空 refresh credential"
+        );
+
+        let configured = format!(
+            "{}{}",
+            origin.replacen("http://", "https://", 1),
+            "/console/topup"
+        );
+        let config = remote_config::RemoteConfig {
+            relay_directory: remote_config::RelayDirectoryPolicy {
+                blocked_hosts: vec![],
+                sites: std::collections::BTreeMap::from([(
+                    "127.0.0.1".to_string(),
+                    remote_config::RelayDirectorySite {
+                        veridrop_host: None,
+                        entry_url: None,
+                        purchase_url: Some(configured.clone()),
+                        usage_url: None,
+                        display_name: None,
+                    },
+                )]),
+            },
+            ..remote_config::RemoteConfig::default()
+        };
+        let purchase_url = remote_config::configured_purchase_url(&config, &op.site_origin)
+            .expect("签名目录解析不该报错")
+            .expect("这个站在目录里配了购买入口");
+
+        // MockRuntime 的 cookies_for_url 恒返回空 ⇒ 走 300ms 启动超时路径（生产 20s）。
+        let window = purchase::purchase_window(relay_id, &op.site_origin);
+        let error = dispatch_site_window(app.handle(), op, window, purchase_url)
+            .await
+            .expect_err("mock 下观察不到轮换，必须按超时收场");
+
+        assert!(
+            error.to_string().contains("重新登录"),
+            "超时错误要有「重新登录」语义：{error}"
+        );
+
+        // 协议隔离：NewAPI 的充值分派不得打任何 sub2api 端点（协议字面量属于
+        // api.rs / 既有测试形状，这里只对照黑名单）。
+        let paths = requests.lock().unwrap().clone();
+        for forbidden in [
+            "/api/v1/settings/public",
+            "/api/v1/user/profile",
+            "/api/v1/auth/refresh",
+        ] {
+            assert!(
+                !paths.iter().any(|path| path == forbidden),
+                "NewAPI 充值分派不该打 sub2api 端点 {forbidden}：{paths:?}"
+            );
+        }
+
+        until_purchase_lease_released(&app, relay_id).await;
+        // 「窗口已销毁」在 MockRuntime 上不可观察：destroy 只清运行时自己的窗口表，
+        // manager 那份（get_webview_window 读的）要等事件循环处理 Destroyed 才清，
+        // 而 mock 的 run_iteration 是 no-op、测试也不驱动 run。销毁动作本身钉在
+        // `newapi_purchase::open` 的 ready-Err 路径（返回前 destroy）与 monitor 兜底；
+        // 这里能观察到的等价不变量是「lease 已还」—— 没有挂着 lease 却无人管理的窗口。
+        assert!(
+            !app.state::<AppState>()
+                .purchase_sessions
+                .is_active(relay_id),
+            "超时收场后 lease 必须已释放"
+        );
+
+        server.abort();
+    }
+
+    #[tokio::test]
+    async fn newapi_purchase_blank_refresh_credential_is_rejected_before_a_window() {
+        let (app, relay_id) =
+            saved_relay_app("https://newapi.example", discovery::BackendKind::NewApi);
+        blank_saved_refresh_credential(&app, relay_id);
+        let op = relay_credentials(&app, relay_id);
+
+        let window = purchase::purchase_window(relay_id, &op.site_origin);
+        let error = dispatch_site_window(
+            app.handle(),
+            op,
+            window,
+            url::Url::parse("https://newapi.example/console/topup").unwrap(),
+        )
+        .await
+        .expect_err("空白 refresh credential 必须在建窗前被拒绝");
+
+        assert!(
+            error.to_string().contains("重新登录"),
+            "错误要指明出路：{error}"
+        );
+        assert!(
+            app.get_webview_window(&purchase::window_label(relay_id))
+                .is_none(),
+            "被拒绝的调用不得留下窗口"
+        );
+        assert!(
+            !app.state::<AppState>()
+                .purchase_sessions
+                .is_active(relay_id),
+            "失败路径的 lease 必须已释放"
+        );
+    }
+
+    #[tokio::test]
+    async fn newapi_purchase_focuses_an_existing_window_without_http_or_lease() {
+        // relay 行故意存 http origin：任何回归（比如有人把续期挪到聚焦检查之前）都会
+        // 真的打上这个哨兵，日志就不再是空的。
+        let (origin, requests, server) = recording_newapi_sentinel().await;
+        let (app, relay_id) = saved_relay_app(&origin, discovery::BackendKind::NewApi);
+
+        // 预建同 label 窗口，模拟「这一行的充值窗已经开着」。
+        let label = purchase::window_label(relay_id);
+        tauri::WebviewWindowBuilder::new(
+            app.handle(),
+            &label,
+            tauri::WebviewUrl::External(url::Url::parse("about:blank").unwrap()),
+        )
+        .build()
+        .expect("预建同 label 窗口");
+
+        let op = relay_credentials(&app, relay_id);
+        let window = purchase::purchase_window(relay_id, &op.site_origin);
+        dispatch_site_window(
+            app.handle(),
+            op,
+            window,
+            url::Url::parse(&format!("{origin}/console/topup")).unwrap(),
+        )
+        .await
+        .expect("聚焦现有窗口是 Ok");
+
+        assert!(
+            requests.lock().unwrap().is_empty(),
+            "同一行第二击不得发任何 HTTP：{:?}",
+            requests.lock().unwrap()
+        );
+        assert!(
+            !app.state::<AppState>()
+                .purchase_sessions
+                .is_active(relay_id),
+            "聚焦路径不得取 lease"
+        );
+
+        server.abort();
+    }
+
+    #[tokio::test]
+    async fn newapi_purchase_ignores_another_relays_lease() {
+        // relay1（NewAPI）的 lease 被占着，relay2（也是 NewAPI）的充值分派照样能拿到
+        // **自己的** lease —— lease 按 relay 键控，一行占用不该拦住另一行。
+        //
+        // 第二腿特意用 NewApi 而不是 sub2api（review F3）：sub2api 分派根本不碰
+        // lease 协调器（跨行隔离对它是平凡成立，走通开窗本身已由 Task 4 的测试覆盖）；
+        // NewApi 腿会真的执行 `try_acquire(relay2)`，被 relay1 的 lease 挡住与否在这里
+        // 才是可观察的。判据：错误是 mock 下的启动超时（「重新登录」）而不是 lease
+        // 占用文案（「正在使用或正在关闭」）—— 后者出现说明 try_acquire 被**别人**
+        // 的 lease 挡了（两条文案都含「充值窗口」，判别要认后者的专属措辞）。
+        let (app, relay1) =
+            saved_relay_app("https://newapi.example", discovery::BackendKind::NewApi);
+        let coordinator = Arc::clone(&app.state::<AppState>().purchase_sessions);
+        let held = coordinator
+            .try_acquire(relay1)
+            .expect("占用 relay1 的 lease");
+
+        // 第二行：另一个 NewAPI 站点账号，自己的 id 与有效 refresh credential。
+        let relay2 = {
+            let state = app.state::<AppState>();
+            let conn = state.db.conn.lock().expect("lock memory database");
+            let id = creds::save_site_with_backend(
+                &conn,
+                "https://newapi2.example",
+                "Second relay",
+                "https://newapi2.example",
+                discovery::BackendKind::NewApi,
+            )
+            .expect("save second relay");
+            creds::save_credentials(
+                &conn,
+                id,
+                creds::AccountIdentity {
+                    id: 8,
+                    label: "Second Account",
+                    login_identifier: "second-account",
+                },
+                "second-access-token",
+                Some("second-refresh-cookie"),
+                None,
+                creds::SessionEnvironment::default(),
+            )
+            .expect("save second credentials");
+            id
+        };
+
+        let op = relay_credentials(&app, relay2);
+        let window = purchase::purchase_window(relay2, &op.site_origin);
+        let error = dispatch_site_window(
+            app.handle(),
+            op,
+            window,
+            url::Url::parse("https://newapi2.example/console/topup").unwrap(),
+        )
+        .await
+        .expect_err("mock 下观察不到轮换，relay2 走自己的超时收场");
+
+        let error = error.to_string();
+        assert!(
+            error.contains("重新登录"),
+            "relay2 拿到了自己的 lease、走进了自己的开窗流程（mock 超时收场）：{error}"
+        );
+        assert!(
+            !error.contains("正在使用或正在关闭"),
+            "出现 lease 占用类错误说明 try_acquire 被别人（relay1）的 lease 挡了：{error}"
+        );
+
+        assert!(
+            coordinator.is_active(relay1),
+            "relay1 的 lease 不受 relay2 的分派影响"
+        );
+        until_purchase_lease_released(&app, relay2).await;
+
+        drop(held);
+    }
+
+    #[tokio::test]
+    async fn newapi_purchase_reports_when_its_own_lease_is_already_held() {
+        let (app, relay_id) =
+            saved_relay_app("https://newapi.example", discovery::BackendKind::NewApi);
+        let op = relay_credentials(&app, relay_id);
+
+        let coordinator = Arc::clone(&app.state::<AppState>().purchase_sessions);
+        let held = coordinator.try_acquire(relay_id).expect("预占自己的 lease");
+
+        let window = purchase::purchase_window(relay_id, &op.site_origin);
+        let error = dispatch_site_window(
+            app.handle(),
+            op,
+            window,
+            url::Url::parse("https://newapi.example/console/topup").unwrap(),
+        )
+        .await
+        .expect_err("自己的 lease 被占时必须明确报错");
+
+        assert!(
+            error.to_string().contains("充值窗口"),
+            "错误要说清是充值窗口占用：{error}"
+        );
+        assert!(
+            app.get_webview_window(&purchase::window_label(relay_id))
+                .is_none(),
+            "不得开第二个窗口"
+        );
+        drop(held);
+    }
+}
